backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai>=0.7.2
edge-tts==6.1.9
python-dotenv==1.0.1
aiofiles==23.2.1
a2wsgi==1.10.7
